Edits & new file
Extensive edits to mallet_tools, including the removal of
Mallet_test_and_write and write_features, as well as the addition of
confusion_tools, a set of functions to create scoreable text files from
gold-standard annotations and tagger output, as well as create and
pretty-print confusion matrices based on those files. The README file
has been appropriately updated to reflect the changes.